WIP moved remaining code from IMFCoreConstraintsChecker into cpl/cc validation classes
