+++ src/main/java/com/netflix/imflibrary/st2067_2/IMFCompositionPlaylist.java
@@ -280,9 +280,6 @@ public class IMFCompositionPlaylist {
             }
         }
 
-        // todo: move call to somewhere else
-        //IMFCoreConstraintsChecker.checkSegments(this, imfErrorLogger);
-
         return virtualTrackMap;
     }
