+++ src/test/java/com/netflix/imflibrary/RESTfulInterfaces/IMPValidatorFunctionalTests.java
@@ -26,7 +26,6 @@ import com.netflix.imflibrary.st0377.header.GenericPackage;
 import com.netflix.imflibrary.st0377.header.Preface;
 import com.netflix.imflibrary.st0377.header.SourcePackage;
 import com.netflix.imflibrary.st2067_2.IMFCompositionPlaylist;
-import com.netflix.imflibrary.st2067_2.IMFCoreConstraintsChecker;
 import com.netflix.imflibrary.utils.ByteArrayByteRangeProvider;
 import com.netflix.imflibrary.utils.ByteArrayDataProvider;
 import com.netflix.imflibrary.utils.ByteProvider;
@@ -270,7 +269,7 @@ public class IMPValidatorFunctionalTests {
         essencesHeaderPartition.add(payloadRecord);
 
         //List<ErrorLogger.ErrorObject> errors = IMPValidator.conformVirtualTracksInCPL(cplPayloadRecord, essencesHeaderPartition);
-        List<ErrorLogger.ErrorObject> errors = IMFCoreConstraintsChecker.checkEssenceDescriptors(imfCompositionPlaylist, essencesHeaderPartition);
+        List<ErrorLogger.ErrorObject> errors = IMPValidator.validateComposition(imfCompositionPlaylist, essencesHeaderPartition);
 
         Assert.assertEquals(errors.size(), 9);
         //The following error occurs because we do not yet support TimedText Virtual Tracks in Photon and the EssenceDescriptor in the EDL corresponds to a TimedText Virtual Track whose entry is commented out in the CPL.
@@ -298,7 +297,7 @@ public class IMPValidatorFunctionalTests {
         PayloadRecord payloadRecord2 = new PayloadRecord(bytes, PayloadRecord.PayloadAssetType.EssencePartition, 0L, resourceByteRangeProvider.getResourceSize());
         List<PayloadRecord> payloadRecords = new ArrayList<PayloadRecord>() {{ add(payloadRecord2); }};
         //errors = IMPValidator.conformVirtualTracksInCPL(cplPayloadRecord, payloadRecords);
-        errors = IMFCoreConstraintsChecker.checkEssenceDescriptors(imfCompositionPlaylist, payloadRecords);
+        errors = IMPValidator.validateComposition(imfCompositionPlaylist, payloadRecords);
         Assert.assertEquals(errors.size(), 4);
 
         inputFile = TestHelper.findResourceByPath("TestIMP/Netflix_Sony_Plugfest_2015/Netflix_Plugfest_Oct2015_ENG51.mxf.hdr");
@@ -307,7 +306,7 @@ public class IMPValidatorFunctionalTests {
         PayloadRecord payloadRecord1 = new PayloadRecord(bytes, PayloadRecord.PayloadAssetType.EssencePartition, 0L, resourceByteRangeProvider.getResourceSize());
         payloadRecords = new ArrayList<PayloadRecord>() {{ add(payloadRecord1); }};
         //errors = IMPValidator.conformVirtualTracksInCPL(cplPayloadRecord, payloadRecords);
-        errors = IMFCoreConstraintsChecker.checkEssenceDescriptors(imfCompositionPlaylist, payloadRecords);
+        errors = IMPValidator.validateComposition(imfCompositionPlaylist, payloadRecords);
         Assert.assertEquals(errors.size(), 4);
 
         inputFile = TestHelper.findResourceByPath("TestIMP/Netflix_Sony_Plugfest_2015/Netflix_Plugfest_Oct2015.mxf.hdr");
@@ -316,7 +315,7 @@ public class IMPValidatorFunctionalTests {
         PayloadRecord payloadRecord0 = new PayloadRecord(bytes, PayloadRecord.PayloadAssetType.EssencePartition, 0L, resourceByteRangeProvider.getResourceSize());
         payloadRecords = new ArrayList<PayloadRecord>() {{ add(payloadRecord0); }};
         //errors = IMPValidator.conformVirtualTracksInCPL(cplPayloadRecord, payloadRecords);
-        errors = IMFCoreConstraintsChecker.checkEssenceDescriptors(imfCompositionPlaylist, payloadRecords);
+        errors = IMPValidator.validateComposition(imfCompositionPlaylist, payloadRecords);
         Assert.assertEquals(errors.size(), 1);
     }
 
@@ -341,7 +340,7 @@ public class IMPValidatorFunctionalTests {
         PayloadRecord payloadRecord2 = new PayloadRecord(bytes, PayloadRecord.PayloadAssetType.EssencePartition, 0L, resourceByteRangeProvider.getResourceSize());
         List<PayloadRecord> payloadRecords = new ArrayList<PayloadRecord>() {{ add(payloadRecord2); }};
         //errors = IMPValidator.conformVirtualTracksInCPL(cplPayloadRecord, payloadRecords);
-        errors = IMFCoreConstraintsChecker.checkEssenceDescriptors(imfCompositionPlaylist, payloadRecords);
+        errors = IMPValidator.validateComposition(imfCompositionPlaylist, payloadRecords);
         Assert.assertEquals(errors.size(), 18);
 
         inputFile = TestHelper.findResourceByPath("TestIMP/NYCbCrLT_3840x2160x23.98x10min/NYCbCrLT_3840x2160x2chx24bitx30.03sec.mxf.hdr");
@@ -349,7 +348,7 @@ public class IMPValidatorFunctionalTests {
         bytes = resourceByteRangeProvider.getByteRangeAsBytes(0, resourceByteRangeProvider.getResourceSize()-1);
         PayloadRecord payloadRecord1 = new PayloadRecord(bytes, PayloadRecord.PayloadAssetType.EssencePartition, 0L, resourceByteRangeProvider.getResourceSize());
         payloadRecords = new ArrayList<PayloadRecord>() {{ add(payloadRecord1); }};
-        errors = IMFCoreConstraintsChecker.checkEssenceDescriptors(imfCompositionPlaylist, payloadRecords);
+        errors = IMPValidator.validateComposition(imfCompositionPlaylist, payloadRecords);
         Assert.assertEquals(errors.size(), 4);
     }
 
@@ -452,7 +451,7 @@ public class IMPValidatorFunctionalTests {
         essencesHeaderPartitionPayloads.add(headerPartition2PayloadRecord);
 
         //List<ErrorLogger.ErrorObject> conformanceErrors = IMPValidator.conformVirtualTracksInCPL(cplPayloadRecord, essencesHeaderPartitionPayloads);
-        List<ErrorLogger.ErrorObject> conformanceErrors = IMFCoreConstraintsChecker.checkEssenceDescriptors(imfCompositionPlaylist, essencesHeaderPartitionPayloads);
+        List<ErrorLogger.ErrorObject> conformanceErrors = IMPValidator.validateComposition(imfCompositionPlaylist, essencesHeaderPartitionPayloads);
 
         Assert.assertEquals(conformanceErrors.size(), 3);
     }
